LibWeb/CSS: Extract interpolation code into its own files
Apart from shrinking StyleComputer a little, we need the ability to get
the current value of a transition from outside of it.

(cherry picked from commit 70d99db9924b068d2a20dd64fec2bd71f7d7b346)
